+++ sentry-android-core/src/main/java/io/sentry/android/core/AndroidTransactionProfiler.java
@@ -11,12 +11,15 @@ import android.os.Process;
 import android.os.SystemClock;
 import io.sentry.HubAdapter;
 import io.sentry.IHub;
+import io.sentry.ILogger;
+import io.sentry.ISentryExecutorService;
 import io.sentry.ITransaction;
 import io.sentry.ITransactionProfiler;
 import io.sentry.PerformanceCollectionData;
 import io.sentry.ProfilingTraceData;
 import io.sentry.ProfilingTransactionData;
 import io.sentry.SentryLevel;
+import io.sentry.SentryOptions;
 import io.sentry.android.core.internal.util.CpuInfoUtils;
 import io.sentry.android.core.internal.util.SentryFrameMetricsCollector;
 import io.sentry.util.Objects;
@@ -28,18 +31,33 @@ import org.jetbrains.annotations.TestOnly;
 
 final class AndroidTransactionProfiler implements ITransactionProfiler {
   private final @NotNull Context context;
-  private final @NotNull SentryAndroidOptions options;
-  private final @NotNull IHub hub;
+  private final @NotNull ILogger logger;
+  private final @Nullable String profilingTracesDirPath;
+  private final boolean isProfilingEnabled;
+  private final int profilingTracesHz;
+  private final @NotNull ISentryExecutorService executorService;
   private final @NotNull BuildInfoProvider buildInfoProvider;
   private boolean isInitialized = false;
   private int transactionsCounter = 0;
   private final @NotNull SentryFrameMetricsCollector frameMetricsCollector;
   private @Nullable ProfilingTransactionData currentProfilingTransactionData;
-
-  private @Nullable ITransaction currentTransaction = null;
   private @Nullable AndroidProfiler profiler = null;
-  private long transactionStartNanos;
+  private long profileStartNanos;
   private long profileStartCpuMillis;
+
+  /**
+   * @deprecated please use a constructor that doesn't takes a {@link IHub} instead, as it would be
+   *     ignored anyway.
+   */
+  @Deprecated
+  public AndroidTransactionProfiler(
+      final @NotNull Context context,
+      final @NotNull SentryAndroidOptions sentryAndroidOptions,
+      final @NotNull BuildInfoProvider buildInfoProvider,
+      final @NotNull SentryFrameMetricsCollector frameMetricsCollector,
+      final @NotNull IHub hub) {
+    this(context, sentryAndroidOptions, buildInfoProvider, frameMetricsCollector);
+  }
 
   public AndroidTransactionProfiler(
       final @NotNull Context context,
@@ -48,25 +66,35 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
       final @NotNull SentryFrameMetricsCollector frameMetricsCollector) {
     this(
         context,
-        sentryAndroidOptions,
         buildInfoProvider,
         frameMetricsCollector,
-        HubAdapter.getInstance());
+        sentryAndroidOptions.getLogger(),
+        sentryAndroidOptions.getProfilingTracesDirPath(),
+        sentryAndroidOptions.isProfilingEnabled(),
+        sentryAndroidOptions.getProfilingTracesHz(),
+        sentryAndroidOptions.getExecutorService());
   }
 
   public AndroidTransactionProfiler(
       final @NotNull Context context,
-      final @NotNull SentryAndroidOptions sentryAndroidOptions,
       final @NotNull BuildInfoProvider buildInfoProvider,
       final @NotNull SentryFrameMetricsCollector frameMetricsCollector,
-      final @NotNull IHub hub) {
+      final @NotNull ILogger logger,
+      final @Nullable String profilingTracesDirPath,
+      final boolean isProfilingEnabled,
+      final int profilingTracesHz,
+      final @NotNull ISentryExecutorService executorService) {
     this.context = Objects.requireNonNull(context, "The application context is required");
-    this.options = Objects.requireNonNull(sentryAndroidOptions, "SentryAndroidOptions is required");
-    this.hub = Objects.requireNonNull(hub, "Hub is required");
+    this.logger = Objects.requireNonNull(logger, "ILogger is required");
     this.frameMetricsCollector =
         Objects.requireNonNull(frameMetricsCollector, "SentryFrameMetricsCollector is required");
     this.buildInfoProvider =
         Objects.requireNonNull(buildInfoProvider, "The BuildInfoProvider is required.");
+    this.profilingTracesDirPath = profilingTracesDirPath;
+    this.isProfilingEnabled = isProfilingEnabled;
+    this.profilingTracesHz = profilingTracesHz;
+    this.executorService =
+        Objects.requireNonNull(executorService, "The ISentryExecutorService is required.");
   }
 
   private void init() {
@@ -75,42 +103,36 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
       return;
     }
     isInitialized = true;
-    final String tracesFilesDirPath = options.getProfilingTracesDirPath();
-    if (!options.isProfilingEnabled()) {
-      options.getLogger().log(SentryLevel.INFO, "Profiling is disabled in options.");
+    if (!isProfilingEnabled) {
+      logger.log(SentryLevel.INFO, "Profiling is disabled in options.");
       return;
     }
-    if (tracesFilesDirPath == null) {
-      options
-          .getLogger()
-          .log(
-              SentryLevel.WARNING,
-              "Disabling profiling because no profiling traces dir path is defined in options.");
+    if (profilingTracesDirPath == null) {
+      logger.log(
+          SentryLevel.WARNING,
+          "Disabling profiling because no profiling traces dir path is defined in options.");
       return;
     }
-    final int intervalHz = options.getProfilingTracesHz();
-    if (intervalHz <= 0) {
-      options
-          .getLogger()
-          .log(
-              SentryLevel.WARNING,
-              "Disabling profiling because trace rate is set to %d",
-              intervalHz);
+    if (profilingTracesHz <= 0) {
+      logger.log(
+          SentryLevel.WARNING,
+          "Disabling profiling because trace rate is set to %d",
+          profilingTracesHz);
       return;
     }
 
     profiler =
         new AndroidProfiler(
-            tracesFilesDirPath,
-            (int) SECONDS.toMicros(1) / intervalHz,
+            profilingTracesDirPath,
+            (int) SECONDS.toMicros(1) / profilingTracesHz,
             frameMetricsCollector,
-            options.getExecutorService(),
-            options.getLogger(),
+            executorService,
+            logger,
             buildInfoProvider);
   }
 
   @Override
-  public synchronized void onTransactionStart(final @NotNull ITransaction transaction) {
+  public synchronized void start() {
     // Debug.startMethodTracingSampling() is only available since Lollipop
     if (buildInfoProvider.getSdkInfoVersion() < Build.VERSION_CODES.LOLLIPOP) return;
 
@@ -119,30 +141,17 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
 
     transactionsCounter++;
     // When the first transaction is starting, we can start profiling
-    if (transactionsCounter == 1) {
-      if (onFirstTransactionStarted(transaction)) {
-        options
-            .getLogger()
-            .log(
-                SentryLevel.DEBUG,
-                "Transaction %s (%s) started and being profiled.",
-                transaction.getName(),
-                transaction.getSpanContext().getTraceId().toString());
-      }
+    if (transactionsCounter == 1 && onFirstStart()) {
+      logger.log(SentryLevel.DEBUG, "Profiler started.");
     } else {
       transactionsCounter--;
-      options
-          .getLogger()
-          .log(
-              SentryLevel.WARNING,
-              "A transaction is already being profiled. Transaction %s (%s) will be ignored.",
-              transaction.getName(),
-              transaction.getSpanContext().getTraceId().toString());
+      logger.log(
+          SentryLevel.WARNING, "A profile is already running. This profile will be ignored.");
     }
   }
 
   @SuppressLint("NewApi")
-  private boolean onFirstTransactionStarted(final @NotNull ITransaction transaction) {
+  private boolean onFirstStart() {
     // init() didn't create profiler, should never happen
     if (profiler == null) {
       return false;
@@ -153,29 +162,43 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
     if (startData == null) {
       return false;
     }
-    transactionStartNanos = startData.startNanos;
+    profileStartNanos = startData.startNanos;
     profileStartCpuMillis = startData.startCpuMillis;
-
-    currentTransaction = transaction;
-
-    currentProfilingTransactionData =
-        new ProfilingTransactionData(transaction, transactionStartNanos, profileStartCpuMillis);
     return true;
+  }
+
+  @Override
+  public synchronized void bindTransaction(final @NotNull ITransaction transaction) {
+    // If the profiler is running, but no profilingTransactionData is set, we bind it here
+    if (transactionsCounter > 0 && currentProfilingTransactionData == null) {
+      currentProfilingTransactionData =
+          new ProfilingTransactionData(transaction, profileStartNanos, profileStartCpuMillis);
+    }
   }
 
   @Override
   public @Nullable synchronized ProfilingTraceData onTransactionFinish(
       final @NotNull ITransaction transaction,
-      final @Nullable List<PerformanceCollectionData> performanceCollectionData) {
+      final @Nullable List<PerformanceCollectionData> performanceCollectionData,
+      final @NotNull SentryOptions options) {
 
-    return onTransactionFinish(transaction, false, performanceCollectionData);
+    return onTransactionFinish(
+        transaction.getName(),
+        transaction.getEventId().toString(),
+        transaction.getSpanContext().getTraceId().toString(),
+        false,
+        performanceCollectionData,
+        options);
   }
 
   @SuppressLint("NewApi")
   private @Nullable synchronized ProfilingTraceData onTransactionFinish(
-      final @NotNull ITransaction transaction,
+      final @NotNull String transactionName,
+      final @NotNull String transactionId,
+      final @NotNull String traceId,
       final boolean isTimeout,
-      final @Nullable List<PerformanceCollectionData> performanceCollectionData) {
+      final @Nullable List<PerformanceCollectionData> performanceCollectionData,
+      final @NotNull SentryOptions options) {
     // check if profiler was created
     if (profiler == null) {
       return null;
@@ -187,15 +210,13 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
 
     // Transaction finished, but it's not in the current profile
     if (currentProfilingTransactionData == null
-        || !currentProfilingTransactionData.getId().equals(transaction.getEventId().toString())) {
+        || !currentProfilingTransactionData.getId().equals(transactionId)) {
       // A transaction is finishing, but it's not profiled. We can skip it
-      options
-          .getLogger()
-          .log(
-              SentryLevel.INFO,
-              "Transaction %s (%s) finished, but was not currently being profiled. Skipping",
-              transaction.getName(),
-              transaction.getSpanContext().getTraceId().toString());
+      logger.log(
+          SentryLevel.INFO,
+          "Transaction %s (%s) finished, but was not currently being profiled. Skipping",
+          transactionName,
+          traceId);
       return null;
     }
 
@@ -203,20 +224,14 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
       transactionsCounter--;
     }
 
-    options
-        .getLogger()
-        .log(
-            SentryLevel.DEBUG,
-            "Transaction %s (%s) finished.",
-            transaction.getName(),
-            transaction.getSpanContext().getTraceId().toString());
+    logger.log(SentryLevel.DEBUG, "Transaction %s (%s) finished.", transactionName, traceId);
 
     if (transactionsCounter != 0) {
       // We notify the data referring to this transaction that it finished
       if (currentProfilingTransactionData != null) {
         currentProfilingTransactionData.notifyFinish(
             SystemClock.elapsedRealtimeNanos(),
-            transactionStartNanos,
+            profileStartNanos,
             Process.getElapsedCpuTime(),
             profileStartCpuMillis);
       }
@@ -230,7 +245,7 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
       return null;
     }
 
-    long transactionDurationNanos = endData.endNanos - transactionStartNanos;
+    long transactionDurationNanos = endData.endNanos - profileStartNanos;
 
     List<ProfilingTransactionData> transactionList = new ArrayList<>(1);
     final ProfilingTransactionData txData = currentProfilingTransactionData;
@@ -240,7 +255,6 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
     currentProfilingTransactionData = null;
     // We clear the counter in case of a timeout
     transactionsCounter = 0;
-    currentTransaction = null;
 
     String totalMem = "0";
     ActivityManager.MemoryInfo memInfo = getMemInfo();
@@ -253,7 +267,7 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
     // Some may not have been really finished, in case of a timeout
     for (ProfilingTransactionData t : transactionList) {
       t.notifyFinish(
-          endData.endNanos, transactionStartNanos, endData.endCpuMillis, profileStartCpuMillis);
+          endData.endNanos, profileStartNanos, endData.endCpuMillis, profileStartCpuMillis);
     }
 
     // cpu max frequencies are read with a lambda because reading files is involved, so it will be
@@ -261,7 +275,9 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
     return new ProfilingTraceData(
         endData.traceFile,
         transactionList,
-        transaction,
+        transactionName,
+        transactionId,
+        traceId,
         Long.toString(transactionDurationNanos),
         buildInfoProvider.getSdkInfoVersion(),
         abis != null && abis.length > 0 ? abis[0] : "",
@@ -281,10 +297,25 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
   }
 
   @Override
+  public boolean isRunning() {
+    return transactionsCounter != 0;
+  }
+
+  @Override
   public void close() {
     // we stop profiling
-    if (currentTransaction != null) {
-      onTransactionFinish(currentTransaction, true, null);
+    if (currentProfilingTransactionData != null) {
+      onTransactionFinish(
+          currentProfilingTransactionData.getName(),
+          currentProfilingTransactionData.getId(),
+          currentProfilingTransactionData.getTraceId(),
+          true,
+          null,
+          HubAdapter.getInstance().getOptions());
+    } else if (transactionsCounter != 0) {
+      // in case the startup profiling is running, and it's not bound to a transaction, we still
+      // stop profiling, but we also have to manually update the counter.
+      transactionsCounter--;
     }
 
     // we have to first stop profiling otherwise we would lost the last profile
@@ -306,17 +337,16 @@ final class AndroidTransactionProfiler implements ITransactionProfiler {
         actManager.getMemoryInfo(memInfo);
         return memInfo;
       }
-      options.getLogger().log(SentryLevel.INFO, "Error getting MemoryInfo.");
+      logger.log(SentryLevel.INFO, "Error getting MemoryInfo.");
       return null;
     } catch (Throwable e) {
-      options.getLogger().log(SentryLevel.ERROR, "Error getting MemoryInfo.", e);
+      logger.log(SentryLevel.ERROR, "Error getting MemoryInfo.", e);
       return null;
     }
   }
 
   @TestOnly
-  @Nullable
-  ITransaction getCurrentTransaction() {
-    return currentTransaction;
+  int getTransactionsCounter() {
+    return transactionsCounter;
   }
 }