+++ sentry/src/main/java/io/sentry/Hub.java
@@ -745,7 +745,14 @@ public final class Hub implements IHub {
       // stop it
       if (samplingDecision.getSampled() && samplingDecision.getProfileSampled()) {
         final ITransactionProfiler transactionProfiler = options.getTransactionProfiler();
-        transactionProfiler.onTransactionStart(transaction);
+        // If the profiler is not running, we start and bind it here.
+        if (!transactionProfiler.isRunning()) {
+          transactionProfiler.start();
+          transactionProfiler.bindTransaction(transaction);
+        } else if (transactionOptions.isStartupTransaction()) {
+          // If the profiler is running and the current transaction is the app startup, we bind it.
+          transactionProfiler.bindTransaction(transaction);
+        }
       }
     }
     if (transactionOptions.isBindToScope()) {