+++ sentry/src/main/java/io/sentry/NoOpTransactionProfiler.java
@@ -15,12 +15,21 @@ public final class NoOpTransactionProfiler implements ITransactionProfiler {
   }
 
   @Override
-  public void onTransactionStart(@NotNull ITransaction transaction) {}
+  public void start() {}
+
+  @Override
+  public boolean isRunning() {
+    return false;
+  }
+
+  @Override
+  public void bindTransaction(@NotNull ITransaction transaction) {}
 
   @Override
   public @Nullable ProfilingTraceData onTransactionFinish(
       @NotNull ITransaction transaction,
-      @Nullable List<PerformanceCollectionData> performanceCollectionData) {
+      @Nullable List<PerformanceCollectionData> performanceCollectionData,
+      @NotNull SentryOptions options) {
     return null;
   }